+++ src/kernel/src/print/ttf_renderer.rs
@@ -63,7 +63,7 @@ impl TtfRenderer {
             initial_offset: 0,
             cache: HashMap::with_capacity(255),
             glyph_cache: HashMap::with_capacity(255),
-            pixel_size: config().kernel.font_size as usize,
+            pixel_size: config().kernel.font_size,
         }
     }
